requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
lxml==5.3.0

# Data Processing
pandas==2.2.2
pyarrow==14.0.1

# Document Parsing
PyPDF2==3.0.1
python-docx==1.1.2

# HTTP Requests
requests==2.31.0

# Environment Variables & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Embeddings & ML
sentence-transformers>=2.2.2
scikit-learn>=1.3.2
torch>=2.1.0
numpy>=1.26.0,<2.0.0
